+++ src/Hooks.cpp
@@ -1,4 +1,4 @@
-#include "UpdateManager.h"
+#include "Managers/UpdateManager.h"
 #include "patches/ArmorRatingScaling.h"
 #include "patches/BashBlockStaminaPatch.h"
 #include "Events.h"
